Refactor JSON submission and handle exceptions
