+++ onadata/apps/api/viewsets/xform_submission_viewset.py
@@ -1,5 +1,4 @@
 import re
-import StringIO
 
 from django.conf import settings
 from django.contrib.auth.models import User
@@ -26,7 +25,7 @@ from onadata.libs.mixins.authenticate_header_mixin import \
     AuthenticateHeaderMixin
 from onadata.libs.mixins.openrosa_headers_mixin import OpenRosaHeadersMixin
 from onadata.libs.renderers.renderers import TemplateXMLRenderer
-from onadata.libs.serializers.data_serializer import SubmissionSerializer
+from onadata.libs.serializers.data_serializer import SubmissionSerializer, JSONSubmissionSerializer
 from onadata.libs.utils.logger_tools import dict2xform, safe_create_instance
 from onadata.apps.api.tools import get_baseviewset_class
 
@@ -36,60 +35,7 @@ BaseViewset = get_baseviewset_class()
 # 10,000,000 bytes
 DEFAULT_CONTENT_LENGTH = getattr(settings, 'DEFAULT_CONTENT_LENGTH', 10000000)
 xml_error_re = re.compile('>(.*)<')
-
-
-def is_json(request):
-    return 'application/json' in request.content_type.lower()
-
-
-def dict_lists2strings(d):
-    """Convert lists in a dict to joined strings.
-
-    :param d: The dict to convert.
-    :returns: The converted dict."""
-    for k, v in d.items():
-        if isinstance(v, list) and all([isinstance(e, basestring) for e in v]):
-            d[k] = ' '.join(v)
-        elif isinstance(v, dict):
-            d[k] = dict_lists2strings(v)
-
-    return d
-
-
-def dict_paths2dict(d):
-    result = {}
-
-    for k, v in d.items():
-        if k.find('/') > 0:
-            parts = k.split('/')
-            if len(parts) > 1:
-                k = parts[0]
-                for p in parts[1:]:
-                    v = {p: v}
-
-        result[k] = v
-
-    return result
-
-
-def create_instance_from_json(username, request):
-    request.accepted_renderer = JSONRenderer()
-    request.accepted_media_type = JSONRenderer.media_type
-    dict_form = request.data
-    submission = dict_form.get('submission')
-
-    if submission is None:
-        # return an error
-        return [_(u"No submission key provided."), None]
-
-    # convert lists in submission dict to joined strings
-    submission_joined = dict_paths2dict(dict_lists2strings(submission))
-    xml_string = dict2xform(submission_joined, dict_form.get('id'))
-
-    xml_file = StringIO.StringIO(xml_string)
-
-    return safe_create_instance(username, xml_file, [], None, request)
-
+    
 
 class XFormSubmissionViewSet(AuthenticateHeaderMixin,
                              OpenRosaHeadersMixin, mixins.CreateModelMixin,
@@ -109,19 +55,28 @@ class XFormSubmissionViewSet(AuthenticateHeaderMixin,
     serializer_class = SubmissionSerializer
     template_name = 'submission.xml'
 
-    def error_response(self, error, is_json_request, request):
-        if not error:
-            error_msg = _(u"Unable to create submission.")
-            status_code = status.HTTP_400_BAD_REQUEST
-        elif isinstance(error, basestring):
-            error_msg = error
-            status_code = status.HTTP_400_BAD_REQUEST
-        elif not is_json_request:
-            return error
-        else:
-            error_msg = xml_error_re.search(error.content).groups()[0]
-            status_code = error.status_code
+    def get_serializer_class(self):
+        if 'application/json' in self.request.content_type.lower():
+            return JSONSubmissionSerializer
 
-        return Response({'error': error_msg},
-                        headers=self.get_openrosa_headers(request),
-                        status=status_code)
+        return SubmissionSerializer
+    
+    def handle_exception(self, exc):
+        if hasattr(exc, 'response'):
+            return exc.response
+        return super(XFormSubmissionViewSet, self).handle_exception(exc)
+        # if not error:
+        #     error_msg = _(u"Unable to create submission.")
+        #     status_code = status.HTTP_400_BAD_REQUEST
+        # elif isinstance(error, basestring):
+        #     error_msg = error
+        #     status_code = status.HTTP_400_BAD_REQUEST
+        # elif not is_json_request:
+        #     return error
+        # else:
+        #     error_msg = xml_error_re.search(error.content).groups()[0]
+        #     status_code = error.status_code
+
+        # return Response({'error': error_msg},
+        #                 headers=self.get_openrosa_headers(request),
+        #                 status=status_code)
